feat: 각 워커 별 비정상 종료 로직 추가 - BackFillStrategy의 backfill 메서드의 책임을 BackFillWorker로 위임하고 CCXT와 DB의 인터페이스 역할을 하도록 변경
